+++ rest_framework_json_api/compat.py
@@ -1,0 +1,4 @@
+try:
+    import collections.abc as collections_abc  # noqa: F401
+except ImportError:
+    import collections as collections_abc  # noqa: F401